sort translation key asc order before import and export
So that each row in translation is consistent on both remote and local

When translation key is not sorted, the order of each row can be random.
It can cause change to a whole sheet or translation file by simply adding a new translation key.

This pr resolve translation key ordering issue by sorting them in ascending order before saving to the translation file or exporting them to google sheet.